Add fastcs-odin and remove epics module
