--- src/ophyd_async/epics/eiger/__init__.py
@@ -1,3 +1,0 @@
-from ._odin_io import Odin, OdinWriter, Writing
-
-__all__ = ["Odin", "OdinWriter", "Writing"]
