+++ src/ophyd_async/fastcs/odin/__init__.py
@@ -1,1 +1,3 @@
-__all__ = []
+from ._odin_io import OdinHdfIO, OdinWriter, OdinWriting
+
+__all__ = ["OdinHdfIO", "OdinWriter", "OdinWriting"]
