+++ tests/fastcs/eiger/test_eiger_detector.py
@@ -4,16 +4,13 @@ import pytest
 
 from ophyd_async.core import DetectorTrigger, init_devices
 from ophyd_async.fastcs.eiger import EigerDetector, EigerTriggerInfo
-from ophyd_async.testing import get_mock_put, set_mock_value
+from ophyd_async.testing import get_mock_put
 
 
 @pytest.fixture
 def detector(RE):
     with init_devices(mock=True):
         detector = EigerDetector("BL03I", MagicMock())
-    set_mock_value(detector.odin.meta_active, "Active")
-    set_mock_value(detector.odin.capture_rbv, "Capturing")
-    set_mock_value(detector.odin.meta_writing, "Writing")
     return detector
